selenium_tests/requirements.txt
# Selenium Tests Requirements
selenium==4.16.0
pytest==7.4.4
pytest-html==4.1.1
pytest-xdist==3.5.0
webdriver-manager==4.0.1
python-dotenv==1.0.0
requests==2.31.0
Pillow==10.2.0
















